+++ doc/changelog.rst
@@ -26,6 +26,27 @@ Bug Fixes
 - Fixes pickling issue with :code:`autode.config.Config` on Windows and in multiprocessing spawn for Linux/macOS
 
 
+1.3.5
+--------
+----------
+
+
+Usability improvements/Changes
+******************************
+*
+
+
+Functionality improvements
+**************************
+-
+
+
+Bug Fixes
+*********
+- Fixes :code:`ERROR` logging level being ignored from environment variable :code:`AUTODE_LOG_LEVEL`
+- Fixes :code:`autode.values.Value` instances generating items with units on division, and throw a warning if multiplying
+
+
 1.3.4
 --------
 ----------